+++ src/components/RandomQuote/index.js
@@ -1,11 +1,31 @@
-import React from "react";
+import React, { Component } from "react";
 import "./style.css";
-/* import API from "../utils/API"; */
+import API from "../../utils/API"
 import { Accordion } from "react-bootstrap";
 import { Button } from "react-bootstrap";
 import { Card } from "react-bootstrap";
 
-function RandomQuote() {
+
+class RandomQuote extends Component {
+	state = {
+	  quote: [],
+	}
+  
+	apiReq = () => {
+	  API.getQuote().then((response) => {
+		  const randNum = Math.floor(Math.random() * response.data.length)
+		  const quote = response.data[randNum].text
+		  console.log(quote)
+		  this.setState({
+			  quote: quote
+		  });
+		  
+	  });
+	}
+	componentDidMount() {
+	  this.apiReq()
+	}
+  render() {
 	return (
 		<Accordion defaultActiveKey="0">
 			<Card>
@@ -17,19 +37,19 @@ function RandomQuote() {
 							data-target="#collapseOne"
 							aria-expanded="false"
 							aria-controls="collapseOne"
-						>
-							GET INSPIRED
+						> GET INSPIRED. 
 						</button>
 					</Accordion.Toggle>
 				</Card.Header>
 				<Accordion.Collapse eventKey="0">
 					<Card.Body className="quote text-center">
-						Insert Motivational Quote Here
+					"{this.state.quote}"
 					</Card.Body>
 				</Accordion.Collapse>
 			</Card>
 		</Accordion>
 	);
+}
 }
 
 export default RandomQuote;
