+++ src/pages/Landing.js
@@ -1,34 +1,15 @@
-import React, { Component } from "react";
+import React from "react";
 import Hero from "../components/Hero";
 import Container from "../components/Container";
 import Row from "../components/Row";
 import Col from "../components/Col";
-import API from "../utils/API";
+import RandomQuote from "../components/RandomQuote"
 
-class Landing extends Component {
-  state = {
-    quote: "",
-  }
 
-  apiReq = () => {
-    API.getQuote().then((response) => {
-        const randNum = Math.floor(Math.random() * response.data.length)
-        const quote = response.data[randNum].text
-        console.log(quote)
-        this.setState({
-            quote: quote
-        });
-        
-    });
-  }
-  componentDidMount() {
-    this.apiReq()
-  }
-render() {
+function Landing () {
 
   return (
     <div>
-      "{this.state.quote}"
       <Hero backgroundImage="https://images.unsplash.com/photo-1609342122563-a43ac8917a3a?ixid=MXwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHw%3D&ixlib=rb-1.2.1&auto=format&fit=crop&w=1050&q=80">
         <h1>Evolve</h1>
         <h2>Connect your mind & body. </h2>
@@ -36,9 +17,16 @@ render() {
       <Container style={{ marginTop: 30 }}>
         <Row>
           <Col size="md-12">
-            <h1>LANDING PAGE</h1>
+            <h1 className="text-center">welcome, user. </h1>
           </Col>
         </Row>
+
+        <Row>
+          <Col size="md-12">
+          <RandomQuote></RandomQuote>
+          </Col>
+        </Row>
+      
         <Row>
           <Col size="md-12">
             <p>
@@ -54,34 +42,10 @@ render() {
               mattis risus ac quam laoreet cursus. Praesent suscipit orci neque, vestibulum
               tincidunt augue tincidunt non. Duis consequat mattis tortor vitae mattis.
             </p>
-            <p>
-              Phasellus at rutrum nisl. Praesent sed massa ut ipsum bibendum porttitor. Sed
-              malesuada molestie velit ac viverra. Quisque a ullamcorper purus. Curabitur luctus mi
-              ac mi hendrerit semper. Nulla tincidunt accumsan lobortis. Mauris convallis sapien non
-              nibh porta accumsan. Nunc volutpat tempus porttitor. Nunc congue dictum egestas.
-              Aliquam blandit mi eu urna scelerisque, vitae volutpat ligula ultricies. Maecenas vel
-              porta augue. Fusce mauris ex, dignissim et lacinia ut, tempus eget nibh.
-            </p>
-            <p>
-              Etiam ut massa efficitur, gravida sapien non, condimentum sapien. Suspendisse massa
-              tortor, facilisis in neque sit amet, scelerisque elementum tortor. Nullam eget nibh
-              sit amet odio lobortis ullamcorper. Nulla bibendum magna nec sem pulvinar lobortis.
-              Mauris et imperdiet urna, vitae lobortis dui. Nunc elementum elit mi, non mattis enim
-              congue at. Proin mi lectus, ullamcorper id hendrerit eu, ultricies vitae lacus. Nunc
-              vehicula, erat eget laoreet condimentum, felis ante malesuada leo, nec efficitur diam
-              nisi eget nisi. Cras arcu lacus, tristique in bibendum vitae, elementum eget lorem.
-              Maecenas vestibulum volutpat orci eu pharetra. Praesent vel blandit ante, nec faucibus
-              libero. Sed ultrices lorem ex, eu facilisis libero convallis ac. Vivamus id dapibus
-              eros. Nullam tempor sem rhoncus porta semper. Proin bibendum vulputate nisl, fringilla
-              interdum elit pulvinar eu. Quisque vitae quam dapibus, vestibulum mauris quis, laoreet
-              massa.
-            </p>
           </Col>
         </Row>
       </Container>
     </div>
-  );
-}
-}
+  )}
 
 export default Landing;
