+++ src/utils/API.js
@@ -1,11 +1,8 @@
 import axios from "axios";
 
 
-//export default {
-  //getWorkouts: function() {
-    //return axios.get("https://dog.ceo/api/breeds/list");
-  //},
+export default {
   getQuote: function() {
     return axios.get("https://type.fit/api/quotes");
   }
-
+}
